+++ Src/AwesomeAssertions/Common/MemberPathSegmentEqualityComparer.cs
@@ -1,4 +1,4 @@
-#if NET6_0_OR_GREATER || NETSTANDARD2_1
+#if NET || NETSTANDARD2_1
 using System;
 #endif
 using System.Collections.Generic;
@@ -42,7 +42,7 @@ internal class MemberPathSegmentEqualityComparer : IEqualityComparer<string>
 
     public int GetHashCode(string obj)
     {
-#if NET6_0_OR_GREATER || NETSTANDARD2_1
+#if NET || NETSTANDARD2_1
         return obj.GetHashCode(StringComparison.Ordinal);
 #else
         return obj.GetHashCode();