+++ build.ps1
@@ -24,7 +24,6 @@ $env:DOTNET_SKIP_FIRST_TIME_EXPERIENCE = 1
 $env:DOTNET_CLI_TELEMETRY_OPTOUT = 1
 $env:DOTNET_MULTILEVEL_LOOKUP = 0
 $env:DOTNET_ROLL_FORWARD = "Major"
-$env:FALLOUT_TELEMETRY_OPTOUT = 1
 $env:DOTNET_CLI_UI_LANGUAGE = "en-US"
 
 ###########################################################################