+++ build.sh
@@ -20,7 +20,6 @@ export DOTNET_CLI_TELEMETRY_OPTOUT=1
 export DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1
 export DOTNET_MULTILEVEL_LOOKUP=0
 export DOTNET_ROLL_FORWARD="Major"
-export FALLOUT_TELEMETRY_OPTOUT=1
 export DOTNET_CLI_UI_LANGUAGE="en-US"
 
 ###########################################################################